Extract printing DAG status from dag_runner into CLI
